Move each hook to own file; move hooks and components to own folders
