+++ client/app/pages/dashboards/DashboardPage.jsx
@@ -24,7 +24,7 @@ import { clientConfig } from "@/services/auth";
 import { policy } from "@/services/policy";
 import { durationHumanize } from "@/lib/utils";
 import PromiseRejectionError from "@/lib/promise-rejection-error";
-import useDashboard, { DashboardStatusEnum } from "./useDashboard";
+import useDashboard, { DashboardStatusEnum } from "./hooks/useDashboard";
 
 import "./DashboardPage.less";
