+++ client/app/pages/dashboards/components/ShareDashboardDialog.jsx
